Reorganised maven modules
- API contains annotations and metadata types used in the generated classes
- Core contains all other classes used at runtime
- Metadata generator uses the API module
- Metadata tests module requires 'core' during the tests to verify class hierachy
- Core tests uses the metadata generator during compilation to have the metadata classes
  generated. This is a reauirement to have a successful compilation

The bottom-side of this configuration is that in Eclipse, we need to launch a new Maven build
to have the metadata re-generated. JDT-APT and m2e-apt will not work with the metadata generator
project opened in the workspace.
Hopefully, re-generating the metadata classes should not happen too often (only when new cases need
to be tested)